Made minor changes to brew_*.sh scripts.
Minor refactoring.

+++ brew_emacs_release.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if pstree -p $PPID | grep -v grep | grep -q -i emacs; then
+    echo "error: you can not run this script from within emacs (because emacs will be removed and installed)."
+    exit 1
+fi
+
 echo "brewing with emacs release"
 brew uninstall --force emacs
 brew install emacs --with-cocoa --with-gnutls --with-imagemagick@6 --with-librsvg --with-mailutils --with-modules
